+++ projects/mini-rx-store/src/lib/feature.ts
@@ -5,25 +5,17 @@ import { createActionTypePrefix, nameUpdateAction, ofType } from './utils';
 import { distinctUntilChanged, map } from 'rxjs/operators';
 import { createFeatureSelector } from './selector';
 
-type SetStateFn<StateType> = (state: StateType) => Partial<StateType>;
-type StateOrCallback<StateType> = Partial<StateType> | SetStateFn<StateType>;
-
 export abstract class Feature<StateType> {
     private readonly actionTypePrefix: string; // E.g. @mini-rx/products
     private readonly actionTypeSetState: string; // E.g. @mini-rx/products/set-state
     private effectCounter = 1; // Used for naming anonymous effects
 
-    protected state$: BehaviorSubject<StateType> = new BehaviorSubject(
-        undefined
-    );
-    private get state(): StateType {
+    protected state$: BehaviorSubject<StateType> = new BehaviorSubject(undefined);
+    get state(): StateType {
         return this.state$.getValue();
     }
 
-    protected constructor(
-        featureName: string,
-        initialState: StateType
-    ) {
+    protected constructor(featureName: string, initialState: StateType) {
         StoreCore.addFeature<StateType>(featureName, initialState);
 
         this.actionTypePrefix = createActionTypePrefix(featureName);
@@ -35,25 +27,25 @@ export abstract class Feature<StateType> {
         StoreCore.select(createFeatureSelector(featureName)).subscribe(this.state$);
     }
 
-    protected setState(
-        stateOrCallback: StateOrCallback<StateType>,
-        name?: string
-    ): void {
+    protected setState(state: Partial<StateType>, name?: string): void {
         StoreCore.dispatch({
-            type: name
-                ? this.actionTypeSetState + '/' + name
-                : this.actionTypeSetState,
-            payload: this.calcNewState(this.state, stateOrCallback),
+            type: name ? this.actionTypeSetState + '/' + name : this.actionTypeSetState,
+            payload: { ...this.state, ...state },
         });
     }
 
-    protected select<K>(mapFn: (state: StateType | AppState) => K, selectFromStore: boolean = false): Observable<K> {
+    protected select<K>(mapFn: (state: StateType) => K, selectFromStore?: boolean): Observable<K>;
+    protected select<K>(mapFn: (state: AppState) => K, selectFromStore?: boolean): Observable<K>;
+    protected select<K, T extends (state: AppState | StateType) => K>(
+        mapFn: T,
+        selectFromStore: boolean = false
+    ): Observable<K> {
         if (selectFromStore) {
             return StoreCore.select(mapFn);
         }
 
         return this.state$.pipe(
-            map((state: StateType) => mapFn(state)),
+            map((state) => mapFn(state)),
             distinctUntilChanged()
         );
     }
@@ -63,13 +55,11 @@ export abstract class Feature<StateType> {
             effectName = this.effectCounter;
             this.effectCounter++;
         }
-        return `${this.actionTypePrefix}/effect/${effectName}`;
+        return `${this.actionTypePrefix}/EFFECT/${effectName}`;
     }
 
     protected createEffect<PayLoadType = any>(
-        effectFn: (
-            payload: Observable<PayLoadType>
-        ) => Observable<StateOrCallback<StateType>>,
+        effectFn: (payload: Observable<PayLoadType>) => Observable<Partial<StateType>>,
         effectName?: string
     ): (payload?: PayLoadType) => void {
         const effectStartActionType = this.getEffectStartActionType(effectName);
@@ -77,10 +67,10 @@ export abstract class Feature<StateType> {
             ofType(effectStartActionType),
             map((action) => action.payload),
             effectFn,
-            map((stateOrCallback) => {
+            map((state: Partial<StateType>) => {
                 return {
                     type: effectStartActionType + '/' + nameUpdateAction,
-                    payload: this.calcNewState(this.state, stateOrCallback),
+                    payload: { ...this.state, ...state },
                 };
             })
         );
@@ -92,22 +82,6 @@ export abstract class Feature<StateType> {
                 type: effectStartActionType,
                 payload,
             });
-        };
-    }
-
-    private calcNewState(
-        state: StateType,
-        stateOrCallback: StateOrCallback<StateType>
-    ): StateType {
-        if (typeof stateOrCallback === 'function') {
-            return {
-                ...state,
-                ...stateOrCallback(state),
-            };
-        }
-        return {
-            ...state,
-            ...stateOrCallback,
         };
     }
 }